% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_clean_table.R
\name{mermaid_clean_table}
\alias{mermaid_clean_table}
\title{Title}
\usage{
mermaid_clean_table(.data, .column_name_case)
}
\arguments{
\item{.data}{Input data}

\item{.column_name_case}{}
}
\description{
Title
}
\examples{
1
}
