% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_map_sites_interactive.R
\name{mermaid_map_sites_interactive}
\alias{mermaid_map_sites_interactive}
\title{Create an interactive map of MERMAID sample events sites}
\usage{
mermaid_map_sites_interactive(
  .data,
  plot_var = NULL,
  basemap = "Esri.WorldImagery",
  jitter = 0.01,
  size = 2,
  colour = "red",
  alpha = 0.5,
  scale = FALSE,
  scale_position = c("bottomright", "bottomleft", "topright", "topleft"),
  legend = TRUE,
  legend_position = c("bottomright", "bottomleft", "topright", "topleft")
)
}
\arguments{
\item{.data}{Data frame with latitude and longitude of sites.}

\item{plot_var}{Variable to plot by (optional).}

\item{basemap}{Basemap layer. See http://leaflet-extras.github.io/leaflet-providers/preview/ for options. Defaults to "Esri.WorldImagery".}

\item{jitter}{Amount of jittering applied to points. Defaults to 0.01. Set to 0 to remove jittering.}

\item{size}{Size of points (when \code{plot_var} is not a numeric variable). Defaults to 2.}

\item{colour}{Colour of points (when \code{plot_var} is not a character, factor, or logical variable). Defaults to red.}

\item{alpha}{Opacity of points. Defaults to 0.5.}

\item{scale}{Whether there should be a scale bar. Defaults to FALSE.}

\item{scale_position}{The position of the scale bar, if there is one. One of "bottomright", "bottomleft", "topright", "topleft". Defaults to "bottomright".}

\item{legend}{Whether there should be a plot legend (if \code{plot_var} is specified). Defaults to TRUE.}

\item{legend_position}{Position of the plot legend. One of "bottomright", "bottomleft", "topright", "topleft". Defaults to "bottomright".}
}
\description{
Create a static map of sites, zoomed in to the sample events sites. If \code{plot_var} is supplied, then the colour of the site points are based on \code{plot_var}. The map returned is a \code{leaflet} plot, and so \code{leaflet} layers can be built upon it for further customization - please see the examples section.
}
\examples{
library(mermaidr)
sample_events <- mermaid_search_projects(name = "XPDC Kei Kecil 2018") \%>\%
  mermaid_get_project_endpoint(endpoint = "beltfishes/sampleevents", limit = 25)

# Default map
mermaid_map_sites_interactive(sample_events)

# Map sites by value of biomass_kgha_avg
mermaid_map_sites_interactive(sample_events, biomass_kgha_avg)

# Map sites by value of reef_exposure
mermaid_map_sites_interactive(sample_events, reef_exposure)

# Remove legend
mermaid_map_sites_interactive(sample_events, reef_exposure, legend = FALSE)

# Change legend position
mermaid_map_sites_interactive(sample_events, reef_exposure, legend_position = "topright")

# Add scale bar
mermaid_map_sites_interactive(sample_events, scale = TRUE)

# Change position of scale bar
mermaid_map_sites_interactive(sample_events, scale = TRUE, scale_position = "topright")

# Since the map returned is a leaflet object, further customization can
# be done with leaflet code. For example, adding a mini-map for further navigation:
mermaid_map_sites_interactive(sample_events) \%>\%
  leaflet::addMiniMap()

# You can also save a static image of the map using the mapview package:
\dontrun{
library(mapview)
m <- mermaid_map_sites_interactive(sample_events, reef_exposure, legend_position = "topright")
mapshot(m, file = "sites_by_management_rule.png")
}
}
