% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_plot_benthic_pit_percent_cover.R
\name{mermaid_plot_benthic_pit_percent_cover}
\alias{mermaid_plot_benthic_pit_percent_cover}
\title{Plot MERMAID benthic PIT percent cover}
\usage{
mermaid_plot_benthic_pit_percent_cover(.data, group_var, compare_var)
}
\arguments{
\item{.data}{Sample event data containing average benthic PIT percent cover, with variables indicated in \code{group_var} and \code{compare_var}}

\item{group_var}{Variable to group by, e.g. \code{percent_cover_by_benthic_category_avg}}

\item{compare_var}{Variable to compare by, e.g. \code{management_rules}}
}
\description{
Plot box plots of benthic PIT percent cover by a grouping variable \code{group_var} (e.g. benthic category), separated by a supplied \code{compare_var}.
}
\examples{
library(mermaidr)
mermaid_search_projects(name = "XPDC Kei Kecil 2018") \%>\%
  mermaid_get_project_endpoint(endpoint = "benthicpits/sampleevents") \%>\%
  mermaid_clean_management_rules() \%>\%
  mermaid_plot_benthic_pit_percent_cover(percent_cover_by_benthic_category_avg, management_rules)
}
