% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_plot_fish_belt_biomass.R
\name{mermaid_plot_fish_belt_biomass}
\alias{mermaid_plot_fish_belt_biomass}
\title{Plot MERMAID fish belt biomass}
\usage{
mermaid_plot_fish_belt_biomass(.data, group_var, compare_var)
}
\arguments{
\item{.data}{Sample event data containing mean fish belt biomass, with variables indicated in \code{group_var} and \code{compare_var}}

\item{group_var}{Variable to group by, e.g. \code{biomass_kgha_by_trophic_group_avg}}

\item{compare_var}{Variable to compare by, e.g. \code{management_rules}}
}
\description{
Plot box plots of fish belt biomass by a grouping variable \code{group_var} (e.g. trophic group), separated by a supplied \code{compare_var}.
}
\examples{
library(mermaidr)
mermaid_search_projects(name = "XPDC Kei Kecil 2018") \%>\%
  mermaid_get_project_endpoint(endpoint = "beltfishes/sampleevents", limit = 100) \%>\%
  mermaid_plot_fish_belt_biomass(biomass_kgha_by_trophic_group_avg, reef_exposure)
}
