% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_map_sites_static.R
\name{mermaid_static_map_bounds}
\alias{mermaid_static_map_bounds}
\title{Get latitude and longitude bounds from a static map}
\usage{
mermaid_static_map_bounds(map)
}
\arguments{
\item{map}{A static map produced from \code{mermaid_map_sites_static}}
}
\value{
A list of length two containing the longitude and latitude bounds of a map.
}
\description{
Get latitude and longitude bounds from a static map
}
\examples{
\donttest{
library(mermaidr)
sample_events <- mermaid_search_projects(name = "XPDC Kei Kecil 2018") \%>\%
  mermaid_get_project_data("fishbelt", "sampleevents", limit = 25)
p <- mermaid_map_sites_static(sample_events)
mermaid_static_map_bounds(p)
}
}
