% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_benthic_category_percent_cover.R
\name{plot_benthic_category_percent_cover}
\alias{plot_benthic_category_percent_cover}
\title{Plot benthic category percent cover}
\usage{
plot_benthic_category_percent_cover(data, variable)
}
\arguments{
\item{data}{Data with percent cover by benthic category, with variables \code{benthic_category} and \code{mean_percent_cover}}

\item{variable}{Variable to compare by, e.g. \code{management_rules}}
}
\description{
Plot box plots of benthic category percent cover, separated by a supplied \code{variable}.
}
\examples{
# plot_benthic_category_percent_cover(fiji_benthicpit_summary, management_rules)
}
