% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{summary_table}
\alias{summary_table}
\title{Benthic PIT and Beltfish Summary Table}
\usage{
summary_table(benthicpit_data, beltfish_data)
}
\arguments{
\item{benthicpit_data}{Benthic PIT data}

\item{beltfish_data}{Beltfish data}
}
\description{
Generates a summary table that includes percent cover for all benthic categories, mean total biomass for all (main) fish families, overall mean total biomass, as well as site metadata, for the given sites:
}
\examples{
# summary_table(fiji_benthicpit_summary, fiji_beltfish_summary)
}
