% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid_boxplot.R
\name{mermaid_plot_benthic_pit_percent_cover}
\alias{mermaid_plot_benthic_pit_percent_cover}
\title{Plot MERMAID Benthic PIT percent cover}
\usage{
mermaid_plot_benthic_pit_percent_cover(
  .data,
  group_var,
  compare_var,
  clean_values = TRUE,
  clean_values_case = c("title", "sentence"),
  replace_dashes = TRUE
)
}
\arguments{
\item{.data}{Sample event data containing mean fish belt biomass, with variables indicated in \code{group_var} and \code{compare_var}. Likely the output from a \code{mermaid_get_project_data()} call.}

\item{group_var}{Variable to group by, e.g. \code{percent_cover_by_benthic_category_avg}}

\item{compare_var}{Variable to compare by, e.g. \code{management_rules}}

\item{clean_values}{Whether to clean the values in \code{group_var} and \code{.compare_var} (with title case as default - see \code{clean_values_case}), in case they are not clean (e.g. contain dashes, underscores, are lowercase, etc). Defaults to TRUE.}

\item{clean_values_case}{The desired clean values case (default is "title"). For example, if a value is "invertivore-mobile", title case converts it to "Invertivore Mobile" while sentence case converts it to "Invertivore mobile". If you don't want to replace dashes with spaces (e.g. get "Interivore-Mobile", set \code{replace_dashes = FALSE}.}

\item{replace_dashes}{Whether to also remove dashes from values when cleaning them. Defaults to TRUE.}
}
\description{
Plot box plots of Benthic PIT percent cover by a grouping variable \code{.group_var} (e.g. benthic category), separated by a supplied \code{compare_var} (e.g. management rules)
}
\examples{
\donttest{
library(mermaidr)
events <- mermaid_search_projects(name = "XPDC Kei Kecil 2018") \%>\%
  mermaid_get_project_data("benthicpit", "sampleevents")

events \%>\%
  mermaid_plot_benthic_pit_percent_cover(percent_cover_by_benthic_category_avg, reef_exposure)
}
}
